Improve traffic class forward progress
Previous traffic class SRB configuration resulted in the majority of the
SRB being shared. If a traffic class had RDMA put operations outstanding
to a down NIC, the majority of the SRB could be consumed. This would
negatively impact bandwidth of other traffic classes.

Solve this problem by increasing the amount of shared bandwidth for RMDA
traffic classes which need to sustain bandwidth.

NETCASSINI-6928

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>