Use BAR 0 size to identify virtual function
The is_physfn and is_virtfn fields of struct pci_dev are only set if
SR-IOV is enabled, and so cannot be relied upon to identify PF vs. VF in
all cases (e.g. non-SR-IOV systems, passthrough PFs or VFs attached to
VMs).

Instead, use the size of BAR 0 (512M for VF, 2G for PF) to indicate
whether the driver is bound to a PF or VF.

Signed-off-by: Richard Halkyard <richard.halkyard@hpe.com>