Replace MSI-X VF-PF mailbox with vsock
The use of reserved fields in the MSI-X table as a mailbox for VF-PF
communication, as described in the CSDG, does not work for the primary
use case of attaching VFs to virtual machines. When a device is passed
through to a virtual machine, the MSI-X table is emulated for security
reasons, and accesses to the reserved fields that we used for
communication are dropped.

To work around this issue, this commit replaces the MSI-X mailbox code
with a similar mechanism communicating over vsock sockets. This has some
limitations and breaks the trivial case of VFs and PF on the same host
(see block comment in cxi/cass_sriov.c for details), but is usable for
the virtualization case in which the MSI-X mailbox approach was not. Due
to this limitation, the existing SR-IOV tests have been configured to be
skipped.